循環参照の回避
